% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HermiteIntegrationFunction.R
\name{HermiteIntegration}
\alias{HermiteIntegration}
\title{Gauss-Hermite quadrature}
\usage{
HermiteIntegration(FUN, sig2 = 1, ...)
}
\arguments{
\item{FUN}{function to be integrated}

\item{sig2}{scalar}

\item{...}{Additional arguments to the function given in FUN}
}
\description{
The one-dimensional integral of a function is numerically approximated with Gauss-Hermite quadrature with 20 nodes. See Abramowitz and Stegun 25.4.46 pp 890 and Table 25.10 pp 924.
}
\details{
Integrals of the form int f(x) phi(x, sigma^2) dx are computed, where phi is the normal distribution density with mean 0 and variance sigma^2. FUN is the function f() and sig2 is sigma^2.
}
\references{
Abramowitz, M. & Stegun, I. A. (Eds.) Handbook of mathematical functions: with formulas, graphs, and mathematical tables. Dover Publications New York, 1972
}
