% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{KDKW_FD_Rcpp}
\alias{KDKW_FD_Rcpp}
\title{Approximate maximum likelihood algorithm based on Kiefer-Wolfowitz stochastic approximation}
\usage{
KDKW_FD_Rcpp(s_obs, theta_0, theta_min, theta_max, simfun,
  fixed_parameters, K, nk, a, ce, alpha = 1, gamma = 1/6, A = 0L,
  C = 0L, use_log = TRUE)
}
\arguments{
\item{s_obs}{Vector of observed summary statistics}

\item{theta_0}{Vector of starting point}

\item{theta_min}{Vector of minimal values for theta}

\item{theta_max}{Vector of maximal values for theta}

\item{simfun}{name of the function which is used to simulate data (further arguments 
to simfun are given with ...)}

\item{K}{number of steps}

\item{nk}{number of simulations per step}

\item{a}{numeric}

\item{ce}{numeric}

\item{alpha}{numeric}

\item{gamma}{numeric}

\item{A}{integer}

\item{C}{integer}

\item{Hfun}{function which computes the bandwidth matrix. Can be one of the functions in bw.r}

\item{Hnum}{how often shall the bandwidth matrix be computed? Possible values: "once" or an integer indicating after how many iterations the bandwidth selection is to be renewed (e. g. 1 => bandwidth selection in each step)}

\item{kernel}{kernel function}

\item{lg}{logical. Use the log-likelihood instead of the likelihood?}
}
\description{
This function approximates the maximum likelihood estimate of the multivariate parameter theta.
It requires a function to simulate data and compute summary statistics at each position of 
the parameter space. These simulations are used to obtain estimates of the likelihood in a 
stochastic approximation algorithm based on finite differences approximation of the gradient 
(Kiefer-Wolfowitz algorithm).
}
\section{Gain sequences}{

The gain sequences are defined following Spall (2013):
\deqn{ a_k = \frac{a}{(k + A)^{\alpha}} \text{ and }
c_k = \frac{c}{(k + C)^{\gamma}}. }{ a_k = a/(k + A)^alpha and c_k = c/(k + C)^gamma. }
Usually, C>0 is only used if the algorithm is resumed. The default values alpha=1 and 
gamma=1/6 are given in Spall (2013).
}

\section{Implementation}{

The function is implemented using Rcpp and RcppArmadillo.
}

\examples{

# bivariate normal distribution

test = KDKW_FD_Rcpp(s_obs = c(1,2), theta_0 = c(0.5,1.5), theta_min = c(-5, -5), 
theta_max = c(5, 5), K = 3000, a = 10, ce = 2, nk = 10, simfun = "Normal",
fixed_parameters = list(VC = diag(2))) 
matplot(test$theta, t="l")

}
\references{
\itemize{
\item Bertl, J.; Ewing, G.; Kosiol, C. & Futschik, A. Approximate Maximum Likelihood Estimation for Population Genetic Inference. Statistical Applications in Genetics and Molecular Biology, 2017, 16, p. 291-312
\item Kiefer, J. & Wolfowitz, J. Stochastic estimation of the maximum of a regression function. The Annals of Mathematical Statistics, 1952, 23, p. 462-466
\item Blum, J. R. Multidimensional stochastic approximation methods. The Annals of Mathematical Statistics, 1954, 25, 737-744
\item Spall, J. C. Introduction to Stochastic Search and Optimization: Estimation, Simulation and Control. Wiley, 2003
}
}
\author{
Johanna Bertl
}
