% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIMcoal.theta.r
\name{SIMcoal.theta}
\alias{SIMcoal.theta}
\title{Simulation function for a simple coalescent model}
\usage{
SIMcoal.theta(nk, theta, n)
}
\arguments{
\item{nk}{Scalar. Number of simulated data sets (trees).}

\item{theta}{Scalar. Scaled mutation rate.}

\item{n}{Scalar. Sample size.}
}
\description{
The function simulates a coalescent tree and mutations and calculates the number of segregating sites under the infinite sites assumption. The number of segregating sites are used as summary statistics. 
The only parameters of the coalescent model are the scaled mutation rate theta and the number of samples, n.
}
\examples{

# Simulating 10 coalescent trees with scaled mutation rate 5 and 20 branches each. The number of segregating sites are: 

SIMcoal.theta(10, 5, 20)

# Estimation with KDKW.FD: 
test2 = KDKW.FD(s.obs = 3, theta.0 = 1, rest = matrix(c(0, 10), nrow=1), simfun = SIMcoal.theta, K = 500, a = 10, ce = 1, nk = 10, n = 5, Hnum = 1, kernel = dnorm2)
plot(test2$theta, t="l")

}
\author{
Johanna Bertl
}
