% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIMCRNpoisson_glmm_regr.R
\name{SIMCRNpoisson_glmm_regr}
\alias{SIMCRNpoisson_glmm_regr}
\alias{SIMCRNpoisson_glmm_beta}
\alias{SIMCRNpoisson_glmm_sigma}
\title{Simulation function for a Poisson model with one random component}
\usage{
SIMCRNpoisson_glmm_regr(nk, theta, seed, x)

SIMCRNpoisson_glmm_beta(nk, theta, seed, sigma, x)

SIMCRNpoisson_glmm_sigma(nk, theta, seed, beta, x)
}
\arguments{
\item{nk}{integer, number of datasets to be simulated}

\item{theta}{numeric, parameter vector. theta = c(beta, sigma)}

\item{seed}{integer. Seed to simulate the random effect and poisson distributed response.}

\item{x}{numeric matrix, explanatory variables}
}
\description{
The function simulates data under a Poisson model with given explanatory variables \code{x} and parameter vector \code{theta}, consisting of the coefficients of the fixed effects, beta, and sigma, the standard deviation of a single random effect with distribution \code{N(0, sigma^2)}. For each of the \code{nk} simulated datasets, a vector of summary statistics is computed, consisting of the regression coefficients estimated in a Poisson model without random effects, and an estimate of the dispersion based on the Pearson statistic (sum of the squared Pearson residuals) of this model. The function is designed to estimate theta by the approximate maximum likelihood algorithm in \code{KDKW.FD} or \code{KDKW.SP}.
}
\details{
The simulations are obtained with the given seed (designed for the use of Common Random Numbers in the Approximate Maximum Likelihood Algorithm).

SIMpoisson_glmm uses \code{glm} to obtain the summary statistics. This probably has a lot of overhead and could maybe be replaced by a faster alternative.
}
\section{Functions}{
\itemize{
\item \code{SIMCRNpoisson_glmm_beta}: SIMCRNpoisson_glmm_beta performs the same simulations as SIMpoisson_glmm, but it is used to estimate beta only when sigma is known.

\item \code{SIMCRNpoisson_glmm_sigma}: SIMCRNpoisson_glmm_sigma performs the same simulations as SIMpoisson_glmm, but it is used to estimate sigma only when beta is known.
}}

\examples{


x = matrix(c(rep(1, 50), rep(0, 100), rep(1, 50)), ncol=2 )
testsim = SIMCRNpoisson_glmm_regr(100, c(3,5,1), seed = 1234, x)
apply(testsim, 2, mean)
testsim2 = SIMCRNpoisson_glmm_regr(100, c(3,5,1), seed = 1234, x)
apply(testsim2, 2, mean)

}
\seealso{
\link{\code{glm}}
}
\author{
Johanna Bertl
}
