% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIMCRNpoisson_glmm_simple.R
\name{SIMCRNpoisson_glmm_simple}
\alias{SIMCRNpoisson_glmm_simple}
\title{Simulation function for a Poisson model with fixed mean and random intercept}
\usage{
SIMCRNpoisson_glmm_simple(nk, theta, seed, n)
}
\arguments{
\item{nk}{integer, number of datasets to be simulated}

\item{theta}{numeric, parameter vector. theta = c(lambda, sigma)}

\item{seed}{integer. Seed to simulate the random effect and poisson distributed response.}

\item{n}{integer, number of observations}
}
\description{
The function simulates data under a Poisson model with fixed mean lambda and a random intercept with with distribution \code{N(0, sigma^2)}. For each of the \code{nk} simulated datasets, a vector of summary statistics is computed, consisting of the mean and an estimate of the dispersion based on the Pearson statistic (sum of the squared Pearson residuals) of this model. The function is designed to estimate theta by the approximate maximum likelihood algorithm in \code{KDKW.FD} or \code{KDKW.SP}.
The samples are obtained with the given seed.
}
\details{
SIMpoisson_glmm_simple uses simGLMM::simDataGLMM.

This function might be replaced by a more flexible function based on simGLMM::simDataGLMMdesign that uses a design matrix.
}
\examples{

SIMCRNpoisson_glmm_simple(1, c(1,0.5), 1234, 10)

set.seed(1234)
dat = simDataGLMM(clus=10, rep=1, fixedMean=1, fixedCoef=NULL, fixedMat=NULL, covM = 0.5, disFam=poisson())
mean(dat$res)

}
\author{
Johanna Bertl
}
