% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIMnormal.mult2a.r
\name{SIMnormal.mult2a}
\alias{SIMnormal.mult2a}
\title{Simulation function for bivariate normal distribution with experimental summary statistics}
\usage{
SIMnormal.mult2a(nk, mu, sigma = c(1, 1), n)
}
\arguments{
\item{nk}{number of simulations}

\item{mu}{2-vector of the mean}

\item{sigma}{2-vector of the variances}

\item{n}{number of observations}
}
\description{
SIMnormal.mult2a simulates data from a bivariate normal distributions with a fixed diagonal variance covariance matrix and computes summary statistics to estimate the mean vector.
}
\details{
Summary statistics: mean of ALL values, product of mean1 and mean2
}
