% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIMnormalmean.anydim.r
\name{SIMnormalmean.anydim}
\alias{SIMnormalmean.anydim}
\title{Simulation function for a multivariate normal distribution}
\usage{
SIMnormalmean.anydim(nk, mu, sigma, n)
}
\arguments{
\item{nk}{number of datasets}

\item{mu}{mean vector}

\item{sigma}{VC matrix}

\item{n}{number of observations}
}
\description{
The function SIMnormalmean.anydim simulates from a multivariate normal distribution and computes the arithmetic means per dimension as summary statistics with the goal to estimate the mean vector. The full variance covariance matrix can be specified (but not estimated).
}
\details{
SIMnormalmean.anydim uses the library mvtnorm.
}
