% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIMpoisson_glmm_regr_ind.R
\name{SIMpoisson_glmm_regr_ind}
\alias{SIMpoisson_glmm_regr_ind}
\title{Simulation function for a Poisson model with one random component}
\usage{
SIMpoisson_glmm_regr_ind(nk, theta, x)
}
\arguments{
\item{nk}{integer, number of datasets to be simulated}

\item{theta}{numeric, parameter vector. theta = c(beta, sigma)}

\item{x}{numeric matrix, explanatory variables}
}
\description{
The function simulates data under a Poisson model with given explanatory variables \code{x} and parameter vector \code{theta}, consisting of the coefficients of the fixed effects, beta, and sigma, the standard deviation of a single random effect with distribution \code{N(0, sigma^2)}. For each of the \code{nk} simulated datasets, a vector of summary statistics is computed, consisting of the regression coefficients estimated in a Poisson model without random effects, and an estimate of the dispersion based on the Pearson statistic (sum of the squared Pearson residuals) of this model. The function is designed to estimate theta by the approximate maximum likelihood algorithm in \code{KDKW.FD} or \code{KDKW.SP}.
}
\details{
SIMpoisson_glmm uses \code{glm} to obtain the summary statistics. This probably has a lot of overhead and could maybe be replaced by a faster alternative.

Note that this is an experimental function that is only for use with x consisting of a set of dummy variables that is equivalent to a single factor that defines groups. The summary statistics are computed separately for each independent group.
}
\examples{


x = matrix(c(rep(1, 50), rep(0, 100), rep(1, 50)), ncol=2 )
testsim = SIMpoisson_glmm_regr_ind(100, c(3,5,1), x)
apply(testsim, 2, mean)

}
\seealso{
\link{\code{glm}}
}
\author{
Johanna Bertl
}
