% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STARTlik.R
\name{STARTlik}
\alias{STARTlik}
\title{Likelihood estimation for starting points}
\usage{
STARTlik(int, sampling, N1, s.obs, simfun, Hfun = bw.nrd0.mult, kernel, nk,
  ...)
}
\arguments{
\item{int}{matrix of intervals (2 columns, p rows)}

\item{sampling}{character. One of "random" or "grid"}

\item{N1}{number of starting points}

\item{s.obs}{observed summary statistics}

\item{simfun}{name of the function which is used for the simulation and computes the summary statistics}

\item{Hfun}{function which computes the bandwidth matrix. Can be one of the functions in bw.R}

\item{kernel}{kernel function for KDE}

\item{nk}{number of simulations for the estimation of the likelihood with kernel density estimation}

\item{...}{further parameters for simfun}
}
\description{
This function draws starting points from a given matrix of intervals (randomly samples from a uniform distribution or on a grid). For each starting point, the likelihood is simulated.
}
\details{
For points on a grid, the number of starting points N1 is approximate. For each dimension, round(N1^(1/p)) equally spaced points in the interval are chosen, with p being the number of dimensions.

Note that this could easily be parallelized. This is currently not implemented, since it depends on the system the program is run on.
}
\examples{

int = matrix(c(-2, 5, 0, 2), ncol=2, nrow=2, byrow=T)

test_start_grid = STARTrandomlik(int, sampling="grid", N1=20, s.obs = c(2.2, 9), simfun = SIMpoisson_glmm, kernel = robust.unscaled.diagonal, nk=20, x = 1:3)

test_start_random = STARTrandomlik(int, sampling="random", N1=20, s.obs = c(2.2, 9), simfun = SIMpoisson_glmm, kernel = robust.unscaled.diagonal, nk=20, x = 1:3)

}
