% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIKpoisson_glmm.R
\name{LIKpoisson_glmm}
\alias{LIKpoisson_glmm}
\title{Likelihood of the Poisson GLMM}
\usage{
LIKpoisson_glmm(beta, sigma, x, y)
}
\arguments{
\item{beta}{scalar}

\item{sigma}{scalar}

\item{x}{vector, regressor}

\item{y}{vector, regressand}
}
\description{
The likelihood of the parameters beta and sigma in a Poisson GLMM is computed. The integral is calculated with Gauss-Hermite quadrature.
}
\examples{

x = 1:3
y = c(2,4,6)
beta = 1
sigma = 1

betavec = seq(0, 5, by =0.1)
likvec = numeric(length(betavec))
for(i in 1:length(likvec)) likvec[i] = LIKpoisson_glmm(betavec[i], sigma, x, y)
plot(betavec, likvec)


}
